Time: 4 ms (21.31%), Space: 47.8 MB (6.57%) - LeetHub

+++ 1018-binary-prefix-divisible-by-5/1018-binary-prefix-divisible-by-5.java
@@ -1,9 +1,9 @@
 class Solution {
     public List<Boolean> prefixesDivBy5(int[] nums) {
         List<Boolean> l=new ArrayList<>();
-        long sum=0;
+        int sum=0;
         for(int i:nums){
-            sum=(long)((sum*2L)+(long)i)%5L;
+            sum=(sum*2+i)%5;
             if(sum%5==0) l.add(true);
             else l.add(false);
         }
